services/session_monitor/requirements.txt
requests==2.31.0
google-generativeai==0.3.2
google-cloud-storage==2.14.0
google-cloud-secret-manager==2.18.0

